QUALITY: Introduce reusable C code adapter
Using C macros will mean we will need to modify the build process
and ensure that the environment variables are added.

C++ should generate only what it needs to generate as should C
generated code. We can reflect this what clib template code is
used in a subclass of the clib translator (ReusableCAdapter)
which adds a new method to check if heap-allocated memory should
be freed.

This means we won't need to remove the calls to free strings
or vectors -- instead the calls in C++ will effectively just
return (see updated templates for vectors and strings).