QUALITY: Remove whitespace in templates
Also update brace placement -- trying to get this consistent with
the other generated code in the existing templates.

Also deprecate '//' comments which are added to the generated code
in preference for ERB '<%# %>' comments, which are not added to the
generated code. These comments are to help the reader of the template
to understand how the mapping works, not for the generated code to
display?