FIX: Fix inconsistent casing for docs translator
We cannot create an array of translators and use that array
as was the case before this fix.

Each time you create a new adapter, you change the casing funcs
for String and Symbol (see define_case_converters in initialize
under AbstractTranslator).

Hence, if you want to run through each adapter, you must
do so in an each loop (i.e., run for each adapter this block
by running the block on all my data then switching to the next
adapter once this block is done). That way we switch to the
new casing functions when we are done with that adapter.